apply_dimension: support changing band names via apply_metadata https://github.com/Open-EO/openeo-geopyspark-driver/issues/1155
